Changed most int fields to double. Reworked of movement input.

+++ src/mainApp/Platform.java
@@ -2,7 +2,7 @@ package mainApp;
 
 public class Platform extends GameObject {
 
-	public Platform(int xCord,int yCord) {
+	public Platform(double xCord,double yCord) {
 		super(xCord, yCord);
 	}
